+++ src/main/java/com/example/animore/Search/SearchHistoryRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Integer> {
+public interface SearchHistoryRepository extends JpaRepository<SearchHistory,Long> {
 
     //최근 검색기록 - 3개
-    List<SearchHistory> findByUserIdxOrderBySearchCreateAtDesc(int userIdx);
+    List<SearchHistory> findByUserIdxOrderBySearchCreateAtDesc(Long userIdx);
 }
